tensorflow_graphics/projects/occupancy_networks/requirements.txt
cython==0.29.2
imageio==2.4.1
numpy==1.15.4
matplotlib==3.0.3
pandas==0.23.4
pillow==7.2.0
scikit-image==0.14.1
scipy==1.1.0
tqdm==4.28.1
trimesh==2.37.7
plyfile==0.7
pyembree==0.1.4
pytest==4.0.2
h5py==2.9.0
pyyaml==5.3.1